Fixed a bug.  MyriaLocalJoin required a columnlist, since it was a subclass of ProjectingJoin.
But there was a OneToOne rule converting Join to MyriaLocalJoin.  This rule failed, since Join has no columnlist.

Created a new rule to convert Join to MyriaLocalJoin, constructing the columnlist that is needed.

Also added a method to class scheme to return a columnlist.

Added some TODOs elsewhere; the format of columnlist may change soon.