datalog/model.py: fix join condition when term contains a constant
In the join: R(1,x),S(x,y), the join condition was being emitted as
R.$0 == S.$0, when it should be looking at R.$1. The problem was that
the join condition was getting the position of the variable `x` by
iterating over a list of a `term.vars()` which didn't include the `1`.
Oops!

Add a TODO in that same code to think about how to handle the case:

R(x), S(x,x). Right now we emit a single join: R.$0 == S.$0, rather than
a compound join R.$0 == S.$1. This is correct -- there is a select
condition on (S.$0 == S.$1), but an optimizer might want to choose
somehow between joining R.$0 on S.$0 or S.$1, etc.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>