raco: run 'reindent' program
I know that Bill prefers two spaces to four, but it's Python standard
and some tools apparently accept it.

For context: I just found a bug trying to switch to using proper
metaclasses for abstract types. The only way I was able to fix it was to
reindent the file in question. WTF, I know, but it worked!

    pip install reindent
    reindent -r raco

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>