sigma-clipping.myl: make it readable
If we're going to put it up on the web, use spacing and alignment that
make sense for that case.

2 spaces, not 4. (Editor is then)

Any clauses (FROM/WHERE/EMIT) that need to span multiple lines should
have the keyword on the left, and arguments intended. One argument per
line.

Combine the tooHigh/tooLow/newBad into one line, since we can now.

Signed-off-by: Daniel Halperin <dhalperi@cs.washington.edu>