PushSelects: tidy push through apply, implement push through groupby
We push selects through groupby if the selects are only on the group
keys (and the group keys are attribute refs). It works the same as
pushing through apply.

The tidying is that now both these rules handle the case when the
Apply/GroupBy contains named OR unnamed references. Before it only
handled unnamed references.

Fix #190