raco: fix bug #240
In the rule RemoveUnusedColumns, I added code to change the
grouping_list and aggregate_list attributes of algebra.GroupBy. There
was a third member variable, column_list, that was defined redundantly
to be grouping_list + aggregate_list.... however, I did not update that
field as well. After the bugfixes of this morning in #229, that
column_list variable was read for the first and only time and thus a bug
was triggered.

Fix this by simply removing that attribute and replacing it with a
function that returns grouping_list + aggregate_list and thus always
keeps the invariant true.