raco-wide: change the interface to compiling/optimizing
We really don't want to pass around (name, op) pairs with implicit
parallel execution semantics. Instead, let's actually instantiate
Store/StoreTemp for the IDBs/Relations we want to define and add a
Parallel, Sequence, DoWhile for the operations we want to run in
parallel, in series, or in a loop. Implement this API change and then
update all the tests and calling code to support it.

TODO: I disabled rather than fixed the Grappa and Clang tests.