myrialang: ensure that some HC dimension always picked
It seems like the minimum workload being sum(child_sizes) might be too
low -- I got an error that the opt_dim_sizes had never been set. Here I
hacked the code to ensure that the first option is always chosen.

@stechu is this the right workaround? If not, what should it be --
maybe a 1-D join as the alternative, for instance?