cfg: inline from back to front
Intuitively, it makes sense to inline back to front:

* Suppose we are able to inline line i into line i+1. After that inline
  operation, we may now be able to inline line i-1 into line i+1. So we
  should check that next. This implies we should move backward in the
  program rather than moving forward.

* Will we lose anything by moving backward? Well, suppose we were able
  to inline line i into line i+1. Could that possibly have impacted the
  ability to inline line i+1 into line i+2? No. Thus, it is not possible
  that going backwards results in more passes over the program than
  going forward.

The only trick to implementation was that when sliding the window
forward, we need to remember line i+1 in the case of a successful
inline.

This is anecdotal, but for the attached chained.myl example, the time to
produce the logical plan was reduced for me from .58 seconds to .26.