raco: support an AppendTemp operator
AppendTemp is like StoreTemp except that it appends tuples. For now, at
least on the Myria backend, we only support Appending to temporary
relations so that the query-level fault tolerance issues are elided.

Also add a rule to Myrialang to map StoreTemp(UnionAll(ScanTemp, ..)) to
an AppendTemp on the final Myria physical plans.