Add recursive idb support (#545)
* use object.id() as the key of object map

* new syntax of do-until-convergence
Each statement in a do-until-convergence is an IDB assignment consists
of: an IDBController, an aggregate, and a MyriaL expression. Add a few
new operators and optimization rules to parse a do-until-convergence
loop into MyriaX operators.

* add MULTIMIN as another aggregate.
MULTIMIN is MIN on multiple columns. The order of the columns is used as
the order of tuple comparison.

* refactor OptimizerTest

* add recursive query tests

* take the option of async_ft
async_ft is set when asynchronous fault-tolerance is used.

* address comments

* fix merge

* fix ftMode

* fixes

* use named attribute refs, change MULTIMIN to LEXMIN

* MyriaConsumer as Unary, add stop_recursion to break infinite loop

* specify join pull order in do-until-convergence

* address comments

* fix flake